refactor(gabinet): extract patients.\$patientId route components (closes #4455)
Split the 2952-line patients detail route into 14 dedicated files under
src/components/gabinet/patients/:

- intake-utils.ts — classifyIntakeItem / groupIntakeSummary helpers
- appointment-row.tsx — shared appointment list row
- patient-sidebar-extra.tsx — statistics + medical notes sidebar
- patient-overview-tab.tsx — overview tab
- patient-appointments-tab.tsx — appointments tab
- patient-beauty-plan-tab.tsx — beauty plan history tab
- patient-payments-tab.tsx — payments tab with action buttons
- patient-billing-tab.tsx — receipts/billing tab
- patient-loyalty-tab.tsx — loyalty points tab
- edit-payment-dialog.tsx — edit payment dialog
- refund-credit-dialog.tsx — refund / request-refund dialog
- add-payment-dialog.tsx — add payment dialog
- gdpr-erase-dialog.tsx — GDPR erase confirmation dialog
- cancel-payment-dialog.tsx — cancel payment dialog

Route file reduced from 2952 → 1215 lines (thin orchestration shell).
All state and handlers remain in the route; components receive them as props,
following the pattern established by the employees route refactor (#4452).
